Update Spotify class to not have static methods and carry out authentication within class itself.

+++ app/Http/Controllers/TrackVibeController.php
@@ -4,63 +4,43 @@ namespace App\Http\Controllers;
 
 use App\Vibe;
 use App\Track;
+
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+
+use App\Spotify\Playlist;
+use App\Spotify\WebAPI;
+
 
 class TrackVibeController extends Controller
 
 {
-	
 
 
-    public function __construct()
-
-    {
-
-        $this->middleware('spotifySession');
-
-        $this->middleware('auth');
-
-        $this->middleware('spotifyAuth');
-
-    }
-
-
-
-
-
-    public function store(Request $request) 
+    public function store(WebAPI $webAPI, Track $track, Playlist $playlist, Vibe $vibe) 
 
     {
 
-    	$trackQuery = Track::where('api_id', request('track-api-id'))->get();
+        if(!$webAPI->userIsAuthorised()) {
+
+            return $webAPI->authorise();
+
+        }
 
 
-    	if ($trackQuery->isEmpty()) {
-
-    		$track = Track::create([
-
-                'api_id' => request('track-api-id')
-
-            ]);
-
-    	} else {
-
-    		$track = $trackQuery[0];
-
-    	}
+        $track = $track->find(request('track-api-id'));
 
 
+    	if (!$track) {
 
-        $this->spotifyAPI()->addPlaylistTracks(request('vibe-api-id'), [
+            $track = Track::create(['api_id' => request('track-api-id')]);
 
-            $track->api_id
-
-        ]);
-
+    	} 
 
 
-        $track->vibes()->attach(request('vibe-id'));
+        $playlist->addTrack($vibe->api_id, $track->api_id);
+
+        $track->vibes()->attach($vibe->id);
 
         return redirect()->back();
 
@@ -71,28 +51,23 @@ class TrackVibeController extends Controller
 
 
 
-    public function destroy(Vibe $vibe, Track $track) 
+    public function destroy(WebAPI $webAPI, Vibe $vibe, Track $track, Playlist $playlist) 
 
     {
+        if(!$webAPI->userIsAuthorised()) {
+
+            return $webAPI->authorise();
+
+        }
+
 
         $vibe->tracks()->detach($track->id);
 
-        $vibeOnSpotify = $this->spotifyAPI()->getPlaylist($vibe->api_id);
-
-
-
-        $tracks = [
-            'tracks' => [
-                ['id' => $track->api_id],
-            ],
-        ];
-
-
-
-        $this->spotifyAPI()->deletePlaylistTracks($vibe->api_id, $tracks, $vibeOnSpotify->snapshot_id);
+        $playlist->deleteTrack($vibe->api_id, $track->api_id);
 
         return redirect()->back();
 
     }
+
 
 }
